Switch to new language view under /lang
