+++ src/app/lang/[id]/loading.tsx
