+++ src/app/lang/_components/lexicon/word-list.tsx
@@ -23,7 +23,7 @@ export async function WordList(props: {
       {words.map((word) => (
         <Link
           key={word.id}
-          href={`/language/${props.conlang.id}/?view=lexicon&word=${word.id}`}
+          href={`/lang/${props.conlang.id}/?view=lexicon&word=${word.id}`}
           className={cn(
             "inline-block w-full max-w-[200px] items-center overflow-x-hidden text-ellipsis rounded-md border border-transparent py-3 text-lg hover:underline md:py-1 md:text-base md:text-muted-foreground",
             props.selectedWordId === word.id
