+++ src/app/lang_old/[id]/_components/lexicon.tsx
