+++ src/app/lang_old/[id]/_components/tags-for-word.tsx
