+++ src/app/lang_old/[id]/_components/word-view.tsx
