--- src/app/language/_components/lexicon/word-view.tsx
@@ -1,27 +1,0 @@
-import { Separator } from "~/components/ui/separator";
-import { getWordById } from "~/server/queries";
-import { WordOwnerView } from "./word-owner-view";
-
-export async function WordView(props: {
-  wordId: number;
-  isConlangOwner: boolean;
-}) {
-  let word;
-  try {
-    word = await getWordById(props.wordId);
-  } catch (error) {
-    console.error("Error:", error);
-    return <div className="py-4 text-center">Error loading word.</div>;
-  }
-  if (props.isConlangOwner) {
-    return <WordOwnerView word={word} />;
-  }
-  return (
-    <div id="word" className="flex flex-col gap-1">
-      <div id="word-header" className="flex items-center gap-2">
-        <h2 className="text-2xl font-medium">{word.text}</h2>
-      </div>
-      <Separator />
-    </div>
-  );
-}
