--- src/app/language/page.tsx
@@ -1,4 +1,0 @@
-export default function LanguagesPage() {
-  // TODO: something
-  return <div>Language Page</div>;
-}
